requirements.txt
anthropic>=0.39.0
pdfplumber>=0.11.0
python-docx>=1.0.0
watchdog>=4.0.0
streamlit>=1.38.0
pyyaml>=6.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit-autorefresh>=0.0.1
